Snapshot: PatternAnalysis: add support for reading snapshot files
To this moment we were just using some files provided in the
configuration file. Now we require those files to be part of snapshot
generated by some variation of DiffKemp build command.

We had to add support for snapshot with LLVM YAML library as it is
currently usually handled by the python interface.