CMake: generate development diffkemp executable
During the development with nix we want to minimize the need to call
pip, as it is not much supported inside of nix.

This executable is adding CMAKE_BINARY_DIR/diffkemp/simpll to the
PYTHONPATH, because then the bindings are working out of the box. As the
`simpll_lib.py` is trying to import local `_simpll` and it going to be
available this way.