CMake: make the simpll library name more idiomatic
Due to the cmake defaults current name of the static simpll library is
`libsimpll-lib.a`. We need different name, because we are creating an
executable named `simpll` which would lead to nameclash.

Hence we can change only the name of the library target result and in
the end the library is going to be named only `libsimpll.a` which is the
presumed name of libraries.