CMake: Tests: always include system libraries in testing executable
The if statement basically has no effect, as if there are some system
libraries, then we get into the statement body and add them to the
build.
Or there are not any and we don't need the specific if as it would
produce an empty string, which would be ignored by the
target_link_libraries macro.