CC WRAPPER: Passing PYTHONPATH to the wrapper
If DiffKemp was not installed and used by running
`build/bin/diffkemp` for creating snapshot of a C project,
the snapshot creation was unsuccessful because the wrapper imports
from the diffkemp package. This commit fixes by passing down to the
subprocess, that calls the wrapper, `PYTHONPATH` containing path to the
DiffKemp package.